requirements.txt
opencv-python
insightface
pinecone-client
numpy
matplotlib
onnxruntime
#onnxruntime-gpu